Configure oauth2 resource server in and remove auth service routing
