--- api-gateway/src/main/java/dev/shrkptv/apigateway/client/AuthFeignClient.java
@@ -1,15 +1,0 @@
-package dev.shrkptv.apigateway.client;
-
-import org.springframework.cloud.openfeign.FeignClient;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestHeader;
-import org.springframework.web.bind.annotation.RequestParam;
-
-@FeignClient(name = "auth-service", url = "${auth-client.url}")
-public interface AuthFeignClient {
-
-    @PostMapping("/validate")
-    String validateToken(@RequestHeader(HttpHeaders.AUTHORIZATION) String token);
-}
