--- api-gateway/src/main/java/dev/shrkptv/apigateway/filter/AuthenticationFilter.java
@@ -1,60 +1,0 @@
-package dev.shrkptv.apigateway.filter;
-
-import dev.shrkptv.apigateway.client.AuthFeignClient;
-import org.springframework.cloud.gateway.filter.GatewayFilter;
-import org.springframework.cloud.gateway.filter.factory.AbstractGatewayFilterFactory;
-import org.springframework.context.annotation.Lazy;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.stereotype.Component;
-import org.springframework.web.server.ResponseStatusException;
-import reactor.core.publisher.Mono;
-
-
-@Component
-public class AuthenticationFilter extends AbstractGatewayFilterFactory<AuthenticationFilter.Config> {
-
-    private final RouteValidator validator;
-    private final AuthFeignClient authFeignClient;
-
-    public AuthenticationFilter(RouteValidator validator, @Lazy AuthFeignClient authFeignClient) {
-        super(Config.class);
-        this.validator = validator;
-        this.authFeignClient = authFeignClient;
-    }
-
-    @Override
-    public GatewayFilter apply(Config config) {
-        return (exchange, chain) -> {
-            if (validator.isSecured.test(exchange.getRequest())) {
-
-                if (!exchange.getRequest().getHeaders().containsKey(HttpHeaders.AUTHORIZATION)) {
-                    return Mono.error(new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Missing Authorization header"));
-                }
-
-                String authHeader = exchange.getRequest().getHeaders().getFirst(HttpHeaders.AUTHORIZATION);
-                if (authHeader != null && authHeader.startsWith("Bearer ")) {
-                    authHeader = authHeader.substring(7);
-                }
-
-                try {
-                    String userEmail = authFeignClient.validateToken(authHeader);
-
-                    ServerHttpRequest modifiedRequest = exchange.getRequest().mutate()
-                            .header("X-User-Email", userEmail)
-                            .build();
-
-                    return chain.filter(exchange.mutate().request(modifiedRequest).build());
-                } catch (Exception e) {
-                    return Mono.error(new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Unauthorized access to application"));
-                }
-            }
-            return chain.filter(exchange);
-        };
-    }
-
-    public static class Config {
-
-    }
-}
