--- api-gateway/src/main/java/dev/shrkptv/apigateway/filter/RouteValidator.java
@@ -1,23 +1,0 @@
-package dev.shrkptv.apigateway.filter;
-
-import org.springframework.http.server.reactive.ServerHttpRequest;
-import org.springframework.stereotype.Component;
-
-import java.util.List;
-import java.util.function.Predicate;
-
-@Component
-public class RouteValidator {
-
-    public static final List<String> openApiEndpoints = List.of(
-            "/api/v1/auth/register",
-            "/api/v1/auth/login",
-            "/api/v1/auth/refresh",
-            "/api/v1/auth/validate"
-    );
-
-    public Predicate<ServerHttpRequest> isSecured =
-            request -> openApiEndpoints
-                    .stream()
-                    .noneMatch(uri -> request.getURI().getPath().contains(uri));
-}
