+++ api-gateway/src/main/resources/application.yml
@@ -4,45 +4,31 @@ server:
 spring:
   application:
     name: 'api-gateway'
+  security:
+    oauth2:
+      resourceserver:
+        jwt:
+          jwk-set-uri: ${JWK_SET_URI}
+          issuer-uri: ${ISSUER_URI}
   cloud:
     gateway:
       server:
         webflux:
           routes:
-            - id: auth-service
-              uri: ${AUTH_SERVICE_URL}
-              predicates:
-                - Path=/api/v1/auth/**
-
             - id: user-service
               uri: ${USER_SERVICE_URL}
               predicates:
                 - Path=/api/v1/users/**
-              filters:
-                - AuthenticationFilter
 
             - id: order-service
               uri: ${ORDER_SERVICE_URL}
               predicates:
                 - Path=/api/v1/orders/**
-              filters:
-                - AuthenticationFilter
 
             - id: payment-service
               uri: ${PAYMENT_SERVICE_URL}
               predicates:
                 - Path=/api/v1/payments/**
-              filters:
-                - AuthenticationFilter
-            - id: auth-service-swagger
-              uri: ${AUTH_SERVICE_URL}
-              predicates:
-                - Path=/v3/api-docs/auth
-              filters:
-                - RewritePath=/v3/api-docs/auth, /v3/api-docs
-              metadata:
-                springdoc:
-                  display-name: "Auth Service API"
 
             - id: user-service-swagger
               uri: ${USER_SERVICE_URL}
@@ -89,9 +75,6 @@ spring:
                   - "*"
                 allow-credentials: true
 
-auth-client:
-  url: ${AUTH_CLIENT_URL}
-
 management:
   endpoint:
     health:
@@ -115,10 +98,8 @@ springdoc:
   swagger-ui:
     path: /swagger-ui.html
     server-url: /
-    urls-primary-name: "Auth Service API"
+    urls-primary-name: "User Service API"
     urls:
-      - name: Auth Service API
-        url: /v3/api-docs/auth
       - name: User Service API
         url: /v3/api-docs/users
       - name: Order Service API
