+++ docker-compose.yml
@@ -251,13 +251,13 @@ services:
     ports:
       - ${API_GATEWAY_PORT}:8080
     environment:
-      AUTH_SERVICE_URL: ${AUTH_SERVICE_URL}
       USER_SERVICE_URL: ${USER_SERVICE_URL}
       ORDER_SERVICE_URL: ${ORDER_SERVICE_URL}
       PAYMENT_SERVICE_URL: ${PAYMENT_SERVICE_URL}
-      AUTH_CLIENT_URL: ${AUTH_CLIENT_URL}
       SERVER_PORT: ${API_GATEWAY_PORT}
       FRONTEND_URL: ${FRONTEND_URL}
+      JWK_SET_URI: ${JWK_SET_URI}
+      ISSUER_URI: ${ISSUER_URI}
     depends_on:
       - auth-service
       - user-service
